moving client code out of integration test.
